# perf: option to improve Date Sorting by pre-parsing date items only once