--- a/packages/excel-export/src/typings/global.d.ts
+++ b/packages/excel-export/src/typings/global.d.ts
@@ -1,1 +1,0 @@
-import 'jest-extended/all';