--- a/test/global.d.ts
+++ b/test/global.d.ts
@@ -1,2 +1,1 @@
-import 'jest-extended/all';
 import 'jsdom-global/register';