@ghiscoding 2024-09-24
@zewa666 so this is what I came up with, it's a decent improvement and all the reasoning is described above. So for 50k, we can see a drop of 15sec to 1.5sec which is quite decent. 

Next step, the pre-parsing could be a bit slow and a potential improvement could be to use Web Workers to do the work in the background which I'll probably take a look at that next. 

Can this be improved further? I guess we could improve it more by using other type of sort (like quick sort or whatever else) **but** I'm not sure that we would be able to use the native browser `.sort()` method and so I think it would overcomplicate the code and I would prefer to stick with the default `.sort()` for that reason. Another side note, most other sort types (quick sort, radix, etc...) are written for number sorting and I'm not even sure it would be possible deviate from the default `.sort()`, so might as well just go with the above approach from above which is to pre-parse all date items.