@zewa666 2024-09-24
but again having those extra steps is acceptable when working with large amounts of data imo.