@ghiscoding 2024-09-24
if you know how we could connect other sort types to the default DataView sort, then I'd be interested to receive a PR for sure. Or did you mean my new approach by "extra steps"?