@zewa666 2024-09-24
yeah I meant your approach with preparing the data for easier/faster sorting